requirements.txt
faiss-cpu==1.12.0
Flask==3.1.2
langchain==0.3.27
langchain-community==0.3.29
langchain-huggingface==0.3.1
nltk==3.9.1
numpy>=1.26.2,<2.0.0
Pillow==11.3.0
protobuf>=3.20.3,<5.0.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.1.1
python-docx==1.2.0
sentence-transformers==5.1.0
tensorflow-cpu>=2.15.0
transformers==4.56.0
upstash-redis==1.4.0
google-generativeai>=0.8.3
gunicorn==21.2.0

